backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "alibabacloud-dm20151123>=1.8.1",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fastapi>=0.121.3",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "playwright>=1.57.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
